feat: Add Planets component for displaying information about Star Wars planets
This commit adds a new component called Planets that is responsible for displaying information about Star Wars planets. The component will be used in the main sections of the application to provide users with details about different planets in the Star Wars universe. This addition enhances the user experience by expanding the content available in the application and providing more comprehensive information about the Star Wars universe.