feat: Update People component to conditionally render EntityDetails based on selectedPerson
This commit updates the People component to conditionally render the EntityDetails component based on the value of the selectedPerson state. If a person is selected, the EntityDetails component will be rendered with the corresponding person's details. Otherwise, the EntityDetails component will not be rendered. This change improves the user experience by dynamically displaying the details of the selected person and hiding the details when no person is selected.